refactor: create service to handle logic
